practical5/requirements.txt
jsonpickle==1.0
pycrypto==2.6.1
secretsharing==0.2.6
